Change parameter type to int in response to michaels pull request.

+++ include/xpp/zmp/zmp_optimizer.h
@@ -160,7 +160,7 @@ private:
    * @param[out] Ek the vector to represent e through a,b,c,d
    * @param[out] non_dependent the influence of e that are not dependent on a,b,c,d
    */
-  void DescribeEByPrev(double k, int dim,
+  void DescribeEByPrev(int k, int dim,
       double start_v, Eigen::VectorXd& Ek, double& non_dependent) const;
 
   /**
@@ -176,7 +176,7 @@ private:
    * @param[out] Ek the vector to represent e through a,b,c,d
    * @param[out] non_dependent the influence of f that are not dependent on a,b,c,d
    */
-  void DescribeFByPrev(double k, int dim,
+  void DescribeFByPrev(int k, int dim,
       double start_v, double start_p, Eigen::VectorXd& Ek, double& non_dependent) const;
 
   template<std::size_t N>
